pr-follow-up: split service bus integration tests by category
