+++ src/Arcus.Messaging.Tests.Core/Correlation/TraceParent.cs
@@ -23,5 +23,14 @@ namespace Arcus.Messaging.Tests.Core.Correlation
 
             return new TraceParent(transactionId, operationParentId);
         }
+
+        public static TraceParent Parse(string diagnosticId)
+        {
+            string[] parts = diagnosticId.Split('-');
+            string transactionId = parts[1];
+            string operationParentId = parts[2];
+
+            return new TraceParent(transactionId, operationParentId);
+        }
     }
 }
