+++ src/Arcus.Messaging.Tests.Integration/Arcus.Messaging.Tests.Integration.csproj
@@ -57,4 +57,8 @@
     <Folder Include="MessagePump\EventHubs\MessageHandling\" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Include="MessagePump\ServiceBusMessagePump.ResiliencyTests.cs" />
+  </ItemGroup>
+
 </Project>
