--- src/Arcus.Messaging.Tests.Workers.ServiceBus/MessageHandlers/OrdersAzureServiceBusAbandonMessageHandler.cs
@@ -1,59 +1,0 @@
-﻿using System.Threading;
-using System.Threading.Tasks;
-using Arcus.EventGrid.Publishing.Interfaces;
-using Arcus.Messaging.Abstractions;
-using Arcus.Messaging.Abstractions.ServiceBus;
-using Arcus.Messaging.Abstractions.ServiceBus.MessageHandling;
-using Arcus.Messaging.Tests.Core.Messages.v1;
-using Azure.Messaging.EventGrid;
-using GuardNet;
-using Microsoft.Extensions.Azure;
-using Microsoft.Extensions.Logging;
-
-namespace Arcus.Messaging.Tests.Workers.MessageHandlers
-{
-    public class OrdersAzureServiceBusAbandonMessageHandler : AzureServiceBusMessageHandler<Order>
-    {
-        private readonly EventGridPublisherClient _eventGridPublisher;
-
-        private static int _deliveryCount;
-
-        public OrdersAzureServiceBusAbandonMessageHandler(
-            IAzureClientFactory<EventGridPublisherClient> clientFactory, 
-            ILogger<OrdersAzureServiceBusAbandonMessageHandler> logger)
-            : base(logger)
-        {
-            Guard.NotNull(clientFactory, nameof(clientFactory));
-
-            _eventGridPublisher = clientFactory.CreateClient("Default");
-        }
-
-        /// <summary>
-        ///     Process a new message that was received
-        /// </summary>
-        /// <param name="order">Message that was received</param>
-        /// <param name="azureMessageContext">Context providing more information concerning the processing</param>
-        /// <param name="correlationInfo">
-        ///     Information concerning correlation of telemetry and processes by using a variety of unique
-        ///     identifiers
-        /// </param>
-        /// <param name="cancellationToken">Cancellation token</param>
-        public override async Task ProcessMessageAsync(
-            Order order,
-            AzureServiceBusMessageContext azureMessageContext,
-            MessageCorrelationInfo correlationInfo,
-            CancellationToken cancellationToken)
-        {
-            if (++_deliveryCount <= 1)
-            {
-                Logger.LogTrace("Abandoning message '{OrderId}'...", order.Id);
-                await AbandonMessageAsync();
-                Logger.LogTrace("Abandoned message '{OrderId}'", order.Id);
-            }
-            else
-            {
-                await _eventGridPublisher.PublishOrderAsync(order, correlationInfo, Logger);
-            }
-        }
-    }
-}
